refactor(components): ゲーム内 UI を VGame.vue から3コンポーネントに分割する
スコア表示・ゲーム終了ダイアログ・パス通知が1ファイルに混在していた。
a11y 対応（#219）で各領域に aria 属性を追加する前に構造を整理するため分割した。
