+++ tests/unit/VGameOver.spec.ts
@@ -5,7 +5,7 @@ import { setActivePinia, createPinia } from "pinia";
 import { createVuetify } from "vuetify";
 import { useGameStore } from "@/stores/game";
 import { CellState } from "@/models/reversi";
-import VGame from "@/components/reversi/VGame.vue";
+import VGameOver from "@/components/reversi/VGameOver.vue";
 
 const mockPush = vi.fn();
 vi.mock("vue-router", () => ({
@@ -17,12 +17,12 @@ vi.mock("canvas-confetti", () => ({ default: mockConfetti }));
 
 const vuetify = createVuetify();
 
-describe("VGame", () => {
+describe("VGameOver", () => {
   beforeEach(() => {
     setActivePinia(createPinia());
     mockConfetti.mockClear();
     mockPush.mockClear();
-    mount(VGame, { global: { plugins: [vuetify] } });
+    mount(VGameOver, { global: { plugins: [vuetify] } });
   });
 
   describe("もう一度ボタン", () => {
@@ -31,7 +31,7 @@ describe("VGame", () => {
       store.board.rows.forEach((row) =>
         row.cells.forEach((cell) => (cell.state = CellState.Black)),
       );
-      mount(VGame, {
+      mount(VGameOver, {
         attachTo: document.body,
         global: { plugins: [vuetify] },
       });
@@ -46,46 +46,12 @@ describe("VGame", () => {
     });
   });
 
-  describe("待ったボタン", () => {
-    it("allowUndo: false のとき待ったボタンが表示されない", () => {
-      const store = useGameStore();
-      store.startGame({ allowUndo: false });
-      const wrapper = mount(VGame, { global: { plugins: [vuetify] } });
-      expect(wrapper.find("[data-testid='undo-button']").exists()).toBe(false);
-    });
-
-    it("allowUndo: true のとき待ったボタンが表示される", () => {
-      const store = useGameStore();
-      store.startGame({ allowUndo: true });
-      const wrapper = mount(VGame, { global: { plugins: [vuetify] } });
-      expect(wrapper.find("[data-testid='undo-button']").exists()).toBe(true);
-    });
-
-    it("allowUndo: true かつ canUndo: false のとき待ったボタンが disabled", () => {
-      const store = useGameStore();
-      store.startGame({ allowUndo: true });
-      const wrapper = mount(VGame, { global: { plugins: [vuetify] } });
-      const btn = wrapper.find("[data-testid='undo-button']");
-      expect(btn.attributes("disabled")).toBeDefined();
-    });
-
-    it("allowUndo: true かつ canUndo: true のとき待ったボタンをクリックで undo が呼ばれる", async () => {
-      const store = useGameStore();
-      store.startGame({ allowUndo: true });
-      store.put(3, 2);
-      const wrapper = mount(VGame, { global: { plugins: [vuetify] } });
-      await wrapper.find("[data-testid='undo-button']").trigger("click");
-      expect(store.canUndo).toBe(false); // undo 後は履歴が空
-    });
-  });
-
   it("ゲームが終了して黒が勝った場合、confetti が呼ばれる", async () => {
     const store = useGameStore();
     store.board.rows.forEach((row) =>
       row.cells.forEach((cell) => (cell.state = CellState.Black)),
     );
     await nextTick();
-
     expect(mockConfetti).toHaveBeenCalledOnce();
   });
 
@@ -95,13 +61,11 @@ describe("VGame", () => {
       row.cells.forEach((cell) => (cell.state = CellState.White)),
     );
     await nextTick();
-
     expect(mockConfetti).toHaveBeenCalledOnce();
   });
 
   it("ゲームが終了していない状態では confetti が呼ばれない", async () => {
     await nextTick();
-
     expect(mockConfetti).not.toHaveBeenCalled();
   });
 
@@ -114,7 +78,6 @@ describe("VGame", () => {
       }),
     );
     await nextTick();
-
     expect(mockConfetti).not.toHaveBeenCalled();
   });
 });
